refactor manage tags and view tagged files dialogs
