+++ src/features/dialogs/ManageTagsDialog.tsx
@@ -1,3 +1,4 @@
+import NiceModal, { useModal } from "@ebay/nice-modal-react"
 import {
     Dialog,
     DialogContent,
@@ -7,30 +8,18 @@ import {
 } from "@src/components/ui/dialog"
 import { ScrollArea } from "@src/components/ui/scroll-area"
 import { Separator } from "@src/components/ui/separator"
+import { useQuery } from "@tanstack/react-query"
+import { Fragment } from "react/jsx-runtime"
 import {
     GET_TOTAL_TAGGED_FILES_MAP_QUERY_KEY,
     getTotalTaggedFilesMap,
-} from "@src/features/files/tags/stats"
-import { TAGS_QUERY_KEY, Tag, getTags } from "@src/features/files/tags/storage"
-import { useQuery } from "@tanstack/react-query"
-import { Fragment } from "react/jsx-runtime"
+} from "../files/tags/stats"
+import { TAGS_QUERY_KEY, getTags } from "../files/tags/storage"
 import { ManageTagCard } from "./ManageTagCard"
 
-interface ManageTagsDialogProps {
-    open: boolean
-    setIsManageTagsDialogOpen: React.Dispatch<React.SetStateAction<boolean>>
-    setActiveTag: React.Dispatch<React.SetStateAction<Tag>>
-    setIsViewTaggedFilesDialogOpen: React.Dispatch<
-        React.SetStateAction<boolean>
-    >
-}
+export default NiceModal.create(() => {
+    const modalHandler = useModal()
 
-export function ManageTagsDialog({
-    setIsManageTagsDialogOpen,
-    setIsViewTaggedFilesDialogOpen,
-    setActiveTag,
-    open,
-}: ManageTagsDialogProps) {
     const { data: tags } = useQuery({
         queryKey: [TAGS_QUERY_KEY],
         queryFn: getTags,
@@ -42,8 +31,8 @@ export function ManageTagsDialog({
     })
 
     return (
-        <Dialog open={open} onOpenChange={setIsManageTagsDialogOpen}>
-            <DialogContent aria-describedby="dialog-description">
+        <Dialog handler={modalHandler}>
+            <DialogContent>
                 <DialogHeader>
                     <DialogTitle>Manage tags</DialogTitle>
                     <DialogDescription id="dialog-description">
@@ -58,13 +47,6 @@ export function ManageTagsDialog({
                             return (
                                 <Fragment key={id}>
                                     <ManageTagCard
-                                        setActiveTag={setActiveTag}
-                                        setIsManageTagsDialogOpen={
-                                            setIsManageTagsDialogOpen
-                                        }
-                                        setIsViewTaggedFilesDialogOpen={
-                                            setIsViewTaggedFilesDialogOpen
-                                        }
                                         tag={tag}
                                         files={stats && stats[id]}
                                     />
@@ -81,4 +63,4 @@ export function ManageTagsDialog({
             </DialogContent>
         </Dialog>
     )
-}
+})
